Improve tuist run CLI output
